Fix Prediction.output_iterator
The `output` variable is shadowed by the yielding loop, so when we hit
the line:

    previous_output = output

the value of `output` refers to the last value of `output` from the loop
above, not the total output we received. That causes the client to
shrink the output down to a single string (which, inconveniently, also
is a valid argument to `len()`), which means we take the wrong slice of
the output the next time through the loop.

Signed-off-by: Dominic Baggott <dominic.baggott@gmail.com>