Consistently return Boolean for delete methods (#359)
Resolves #356 

When deleting a model, Replicate's API returns with a 204 and an empty
response body. The client library fails to decode this and raises and
exception.

This PR updates the `replicate.models.delete` method to return `bool`
instead to indicate success. It also harmonizes the other delete methods
for deployments and files to do the same.

Signed-off-by: Mattt Zmuda <mattt@replicate.com>