Fix a couple of bugs in the base64 file_encoding_strategy (#398)
This commit adds tests for the `file_encoding_strategy` argument for
`replicate.run()` and fixes two bugs that surfaced:

1. `replicate.run()` would convert the file provided into base64 encoded
data but not a valid data URL. We now use the `base64_encode_file`
function used for outputs.
2. `replicate.async_run()` accepted but did not use the
`file_encoding_strategy` flag at all. This is fixed, though it is worth
noting that `base64_encode_file` is not optimized for async workflows
and will block. This might be okay as the file sizes expected for data
URL payloads should be very small.