Dereference OpenAPI schema returned by Function.openapi_schema (#439)
This PR updates the `openapi_schema` method in `Function` and
`AsyncFunction` to return a dereferenced API schema. This fixes a bug
with transformation for models that output objects with named
interfaces.

This PR also introduces a breaking change in that
`Function.openapi_schema()` is now a method rather than a property to
match `AsyncFunction.openapi_schema()` the latter needs to be a method
because the lookup of the version is both lazy and async.

Lastly, the PR fixes an issue with the implementation where the latest
OpenAPI schema was always returned. Now we return the version specified
in the ref (or the latest if unspecified).