Fix dereferencing in Function and AsyncFunction
We were accidentally stripping out all components in the schemas due to
the Input/Output being referenced by PredictionRequest and
PredictionResponse and those in turn being referenced by the various
`path` entries. This commit ensures we retain Input and Output and
cleans up the object to only contain relevant fields.